1. Encontrar el empleado con mayor salario. [max()] 2. Encontrar el empleado con menor salario. [min()] 3. Ordenar los empleados por nombre. [sorted()] 4. Hallar la suma de los salarios de todos los empleados cuyo salario es mayor a 700000. [filter(), map() y reduce()] 5. Determinar el número total de empleados cuyo apellido comienza por la letra ‘A’ o ‘a’. [filter() y count()] 6. Los 5 primeros empleados con el mayor salario. [sorted() y limit()]

+++ src/main/java/com/dowesoft/herencia/Empleado.java
@@ -18,7 +18,7 @@ public class Empleado extends Persona{
 
     @Override
     public String toString() {
-        return this.getId()+ ") " + this.getNombre()+ " " + this.getApellido()+ " $"+ this.getSalario()+ " "; //To change body of generated methods, choose Tools | Templates.
+        return this.getId()+ ") " + this.getNombre()+ " " + this.getApellido()+ " $"+ ((Double)this.getSalario()).intValue()+ " "; //To change body of generated methods, choose Tools | Templates.
     }
